Renaming sol-warnings to static-analysis
