+++ src/app/static-analysis.js
