+++ test/index.js
@@ -1,4 +1,4 @@
 require('./compiler-test');
 require('./gist-handler-test');
 require('./query-params-test');
-require('./sol-warnings-test');
+require('./static-analysis-test');
